+++ projects/oblique-galaxy/src/lib/components/oblique-galaxy/obg-oblique-galaxy.component.ts
@@ -3,7 +3,7 @@ import {MatButton} from '@angular/material/button';
 import {ObButtonModule} from '@oblique/oblique';
 
 @Component({
-	selector: 'lib-oblique-galaxy',
+	selector: 'obg-oblique-galaxy',
 	standalone: true,
 	imports: [MatButton, ObButtonModule],
 	// eslint-disable-next-line @angular-eslint/component-max-inline-declarations
@@ -12,4 +12,4 @@ import {ObButtonModule} from '@oblique/oblique';
 		<button mat-button obButton="secondary" type="button">Secondary</button>
 		<button mat-button obButton="tertiary" type="button">Tertiary</button>`
 })
-export class ObliqueGalaxyComponent {}
+export class ObgObliqueGalaxyComponent {}
