+++ projects/oblique-galaxy/src/lib/components/oblique-galaxy/oblique-galaxy.component.stories.ts
@@ -1,4 +1,4 @@
-import {ObliqueGalaxyComponent} from './oblique-galaxy.component';
+import {ObgObliqueGalaxyComponent} from './obg-oblique-galaxy.component';
 import {Meta, StoryObj, moduleMetadata} from '@storybook/angular';
 import {HttpClientModule} from '@angular/common/http';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
@@ -6,9 +6,9 @@ import {MatButtonModule} from '@angular/material/button';
 import {ObliqueModule} from '@oblique/oblique'; // More on how to set up stories at: https://storybook.js.org/docs/writing-stories
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories
-const meta: Meta<ObliqueGalaxyComponent> = {
+const meta: Meta<ObgObliqueGalaxyComponent> = {
 	title: 'Example/Galaxy',
-	component: ObliqueGalaxyComponent,
+	component: ObgObliqueGalaxyComponent,
 	tags: ['autodocs'],
 	decorators: [
 		moduleMetadata({
@@ -18,7 +18,7 @@ const meta: Meta<ObliqueGalaxyComponent> = {
 } as Meta;
 
 export default meta;
-type Story = StoryObj<ObliqueGalaxyComponent>;
+type Story = StoryObj<ObgObliqueGalaxyComponent>;
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export const Default: Story = {};
